fix: 회원 정보 수정 시 필드 Blank 체크 추가, 기본 프로필 이미지 기능 추가

+++ module-presentation/src/main/java/com/depromeet/image/facade/ImageFacade.java
@@ -83,9 +83,11 @@ public class ImageFacade {
 
     public void changeProfileImageUrl(Long memberId, String imageName) {
         // 존재하는 이미지를 지운다
-        Member member = memberService.findById(memberId);
-        if (member.getProfileImageUrl() != null && !member.getProfileImageUrl().isEmpty()) {
-            imageDeleteUseCase.deleteProfileImage(member.getProfileImageUrl());
+        if (!isDefaultProfileNumber(imageName)) {
+            Member member = memberService.findById(memberId);
+            if (member.getProfileImageUrl() != null && !member.getProfileImageUrl().isEmpty()) {
+                imageDeleteUseCase.deleteProfileImage(member.getProfileImageUrl());
+            }
         }
         // 멤버의 프로필 이미지 URL 정보를 수정한다
         memberService.updateProfileImageUrl(memberId, imageName);
@@ -98,5 +100,11 @@ public class ImageFacade {
 
     public void deleteAllImagesByMemoryId(Long memoryId) {
         imageDeleteUseCase.deleteAllImagesByMemoryId(memoryId);
+    }
+
+    private static final List<String> DEFAULT_IMAGE_NAMES = List.of("1", "2", "3", "4");
+
+    private static boolean isDefaultProfileNumber(String imageName) {
+        return DEFAULT_IMAGE_NAMES.contains(imageName);
     }
 }
